% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang_vec.R
\name{ang.vec}
\alias{ang.vec}
\title{Angle between vectors}
\usage{
ang.vec(v,w)
}
\arguments{
\item{v}{Vector in R3}

\item{w}{Vector in R3}
}
\description{
This function determines the angle between two vectors v and w.
}
\keyword{angle}
\keyword{vector}
