% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chamber_angles.R
\name{chamber.angles}
\alias{chamber.angles}
\title{Angle between chambers}
\usage{
chamber.angles(X, Y, Z)
}
\arguments{
\item{X}{x coordinates of all chamber centroids}

\item{Y}{y coordinates of all chamber centroids}

\item{Z}{z coordinates of all chamber centroids}
}
\description{
Calculates the angle each chamber makes with its previous two chambers.
}
\keyword{angle}
\keyword{centroid}
\keyword{chamber}
