% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chamber_distances.R
\name{chamber.distances}
\alias{chamber.distances}
\title{Distance between all chambers of a foram}
\usage{
chamber.distances(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}
}
\description{
This function creates a matrix with distances between every possible pair of chambers in R^3 based on the chamber centroid coordinates
}
\keyword{centroid}
\keyword{chamber}
