% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chambers_in_whorl.R
\name{chambers.in.whorl}
\alias{chambers.in.whorl}
\title{Chambers in the whorl}
\usage{
chambers.in.whorl(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}
}
\description{
This function calculates the number of chambers in the final whorl, at the time that chamber i was built.
}
\keyword{centroid}
\keyword{chamber}
