% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chamber_order.R
\name{check.chamber.order}
\alias{check.chamber.order}
\title{Check if chambers are in the correct order}
\usage{
check.chamber.order(X, Y, Z, proloculus=TRUE)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}

\item{proloculus}{Indicates whether the proloculus is present. Defaults to TRUE}
}
\description{
This function checks if the chambers of a foram are in the correct order by analysing the
angles between consecutive chambers. In species with three or more chambers in a whorl
the angles between consecutive chambers are expected to be larger than 60 degrees. The
exception is the angle between the first three chambers, which can be smaller than 60 degrees. In
specimens with the proloculus known to be present the first angle is not taken into
consideration. If the proloculus is not present due to for example dissolution, the first
angle can be taken into account.
}
\keyword{chamber}
\keyword{whorl}
