% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coiling_direction.R
\name{coiling.direction}
\alias{coiling.direction}
\title{Coiling direction}
\usage{
coiling.direction(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}
}
\description{
This function determines coiling direction at the time chamber i was built, relative to
the two previous chambers.
}
\keyword{centroid}
\keyword{chamber}
\keyword{coiling}
