% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prod.R
\name{cross.prod}
\alias{cross.prod}
\title{Cross product}
\usage{
cross.prod(P, Q, R)
}
\arguments{
\item{P}{xyz coordinates of chamber i}

\item{Q}{xyz coordinates of chamber i+1}

\item{R}{xyz coordinates of chamber i+2}
}
\description{
Calculates a vector orthogonal to the plane spanned by three consequtive chamber centroids.
Growth direction determines the direction of the orthogonal vector (i.e. up or down) so
centroids are given in the direction of growth, with P being the oldest chamber of the three,
Q the subsequent chamber and R the subsequent chamber after that.
}
\keyword{centroid}
\keyword{chamber}
