% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram_growth_3D.R
\name{foram.growth.3D}
\alias{foram.growth.3D}
\title{Three-dimensional foram growth}
\usage{
foram.growth.3D(n, X, Y, Z, proloculus=TRUE)
}
\arguments{
\item{n}{chamber number}

\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}

\item{proloculus}{Indicates whether the proloculus is present in the dataset. Defaults to TRUE}
}
\description{
This function calculates the distances between all chambers, angles between subsequent chambers,
trochospirality of each chamber, the number of chambers in the final whorl at the time
each chamber was built and coiling direction at every step in the ontogeny. It orders the data
by increasing chamber number first.
}
\keyword{angle}
\keyword{centroid}
\keyword{chamber}
\keyword{distance}
\keyword{trochospirality}
\keyword{whorl}
