% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_whorls.R
\name{total.whorls}
\alias{total.whorls}
\title{Total number of whorls}
\usage{
total.whorls(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}
}
\description{
This function calculates the number of complete whorls plus any remaining chambers per foram. Returns as
c(number of complete whorls, number of remaining chambers).
}
\keyword{centroid}
\keyword{chamber}
\keyword{whorl}
