% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trochospirality.R
\name{trochospirality}
\alias{trochospirality}
\title{Trochospirality}
\usage{
trochospirality(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chambers}

\item{Y}{Vector with Y-coordinates of all chambers}

\item{Z}{Vector with Z-coordinates of all chambers}
}
\description{
This function calculates the angle a chamber makes with the plane defined by the centroids of the previous three
chambers.
}
\keyword{centroid}
\keyword{chamber}
\keyword{trochospirality}
