% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsacculifer.R
\docType{data}
\name{Tsacculifer}
\alias{Tsacculifer}
\title{Example data to showcase the foram3D package functions}
\format{
A dataframe with 5 columns and 22 rows
\describe{
\item{Chamber}{Chamber number}
\item{Volume}{Chamber volume}
\item{CentroidX}{X-coordinate of chamber centroid}
\item{CentroidY}{Y-coordinate of chamber centroid}
\item{CentroidZ}{Z-coordinate of chamber centroid}
}
}
\usage{
Tsacculifer
}
\description{
Contains chamber number, volume and centroid data of a
Trilobatus sacculifer specimen
}
\examples{
data(Tsacculfier)
}
\keyword{datasets}
