% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whorl.R
\name{whorl}
\alias{whorl}
\title{Whorl number}
\usage{
whorl(X,Y,Z)
}
\arguments{
\item{X}{Vector with X-coordinates of all chamber centroids}

\item{Y}{Vector with Y-coordinates of all chamber centroids}

\item{Z}{Vector with Z-coordinates of all chamber centroids}
}
\description{
Determines the whorl number of each chamber
}
\keyword{centroid}
\keyword{chamber}
\keyword{whorl}
